+++ system-tests/runtimes/file-transfer-consumer/src/main/java/org/eclipse/dataspaceconnector/extensions/api/ApiEndpointExtension.java
@@ -21,6 +21,14 @@ import org.eclipse.dataspaceconnector.spi.system.ServiceExtension;
 import org.eclipse.dataspaceconnector.spi.system.ServiceExtensionContext;
 import org.eclipse.dataspaceconnector.spi.transfer.TransferProcessManager;
 import org.eclipse.dataspaceconnector.spi.transfer.store.TransferProcessStore;
+import org.eclipse.dataspaceconnector.spi.types.domain.transfer.ProvisionedResource;
+import org.eclipse.dataspaceconnector.spi.types.domain.transfer.StatusChecker;
+import org.eclipse.dataspaceconnector.spi.types.domain.transfer.StatusCheckerRegistry;
+import org.eclipse.dataspaceconnector.spi.types.domain.transfer.TransferProcess;
+
+import java.io.File;
+import java.nio.file.Path;
+import java.util.List;
 
 public class ApiEndpointExtension implements ServiceExtension {
     @Inject
@@ -35,6 +43,9 @@ public class ApiEndpointExtension implements ServiceExtension {
     @Inject
     TransferProcessStore transferProcessStore;
 
+    @Inject
+    StatusCheckerRegistry statusCheckerRegistry;
+
     @Override
     public String name() {
         return "API Endpoint";
@@ -43,5 +54,9 @@ public class ApiEndpointExtension implements ServiceExtension {
     @Override
     public void initialize(ServiceExtensionContext context) {
         webService.registerResource(new ConsumerApiController(context.getMonitor(), processManager, negotiationManager, transferProcessStore));
+        statusCheckerRegistry.register("File", (transferProcess, resources) -> {
+            var file = new File(transferProcess.getDataRequest().getDataDestination().getProperty("path"));
+            return file.exists() && file.isDirectory() && file.listFiles().length > 0;
+        });
     }
 }