+++ CHANGELOG.md
@@ -30,8 +30,9 @@ the detailed section referring to by linking pull requests or issues.
 * Refactor ids token validation as extension (#625)
 * All `CosmosDocument` subclasses now use a configurable partition key (#780)
 * Add `findAll` method to `TransferProcessStore` (#859)
-* Enabled pluggable transfer service in DPF (#844)
 * Add data-management api to the samples (#733)
+* Enable pluggable transfer service in DPF (#844)
+* Apply 2-state transition pattern to `TransferProcessManager` (#831)
 
 #### Removed